testdrive: add postgres-verify-slot action
Adds a new action to query Postgres' pg_replication_slots table in
order to confirm that a slot exists or does not exist. This is
required to test that we clean up the upstream replication slot when a
user DROPs a Postgres source.
